struct: add providers folder , move GlobalProviders to providers folder
